Changing design of user log in pages.
